+++ tutorials/FaceNormals/FaceNormals.cpp
@@ -60,10 +60,10 @@ int main()
     }
 
     // copy faces
-    for (int i = 0; i < srcMesh.F.size(); ++i) {
+    for (int i = 0; i < (int)srcMesh.F.size(); ++i) {
         const std::vector<int>& f = srcMesh.F[i];
         assert(f.size() == 3); // we assume triangle meshes for simplicity
-        for (int j = 0; j < f.size(); ++j) {
+        for (int j = 0; j < (int)f.size(); ++j) {
             srcMesh.faceIndicesArray.push_back(f[j]);
         }
 
@@ -84,7 +84,7 @@ int main()
     }
 
     // copy vertices
-    for (int i = 0; i < cutMesh.V.size(); ++i) {
+    for (int i = 0; i < (int)cutMesh.V.size(); ++i) {
         const std::vector<double>& v = cutMesh.V[i];
         assert(v.size() == 3);
         cutMesh.vertexCoordsArray.push_back(v[0]);
@@ -93,10 +93,10 @@ int main()
     }
 
     // copy faces
-    for (int i = 0; i < cutMesh.F.size(); ++i) {
+    for (int i = 0; i < (int)cutMesh.F.size(); ++i) {
         const std::vector<int>& f = cutMesh.F[i];
         assert(f.size() == 3);
-        for (int j = 0; j < f.size(); ++j) {
+        for (int j = 0; j < (int)f.size(); ++j) {
             cutMesh.faceIndicesArray.push_back(f[j]);
         }
 
@@ -498,7 +498,7 @@ int main()
 
         // write normals
 
-        for (int p = 0; p < ccNormals.size(); ++p) {
+        for (int p = 0; p < (int)ccNormals.size(); ++p) {
             Eigen::Vector3d n = ccNormals[p];
             file << "vn " << std::setprecision(std::numeric_limits<long double>::digits10 + 1) << n.x() << " " << n.y() << " " << n.z() << std::endl;
         }
